requirements.txt
flask
flask-cors
gunicorn
playwright
beautifulsoup4
google-genai
